% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_redcap_array}
\alias{assert_redcap_array}
\title{Check if argument is of type redcap_array}
\usage{
assert_redcap_array(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if argument is of type redcap_array
}
