% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapr.R
\name{edit_credential_local}
\alias{edit_credential_local}
\title{Edit tokens and other credentials in a file}
\usage{
edit_credential_local(path_credential = "~/.REDCapR")
}
\arguments{
\item{path_credential}{The file path to the CSV containing the credentials.
Required.}
}
\value{
Target path, invisibly.
}
\description{
Edit tokens and other credentials in a file
}
