% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_metadata}
\alias{export_metadata}
\title{Export Metadata (Data Dictionary)}
\usage{
export_metadata(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  format = c("xml", "csv", "json"),
  fields = NULL,
  forms = NULL,
  return_format = c("xml", "csv", "json")
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{format}{csv, json, xml \verb{[default]}}

\item{fields}{an array of field names specifying specific fields you wish to
pull (by default, all metadata is pulled)}

\item{forms}{an array of form names specifying specific data collection
instruments for which you wish to pull metadata (by default, all metadata is
pulled). NOTE: These 'forms' are not the form label values that are seen on
the webpages, but instead they are the unique form names seen in Column B of
the data dictionary.}

\item{return_format}{csv, json, xml - specifies the format of error messages.
If you do not pass in this flag, it will select the default format for you
passed based on the 'format' flag you passed in or if no format flag was
passed in, it will default to 'xml'.}
}
\value{
\link[httr:response]{httr::response} object containing Metadata from the project (i.e.
Data Dictionary values) in the format specified ordered by the field order.
}
\description{
This method allows you to export the metadata for a project.
}
\note{
To use this method, you must have API Export privileges in the project.
}
\examples{
\dontrun{
export_metadata(token = my_token, format = "csv")
}
}
