% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_setdiff}
\alias{file_setdiff}
\alias{file_setdiff.wucap_file_keys}
\alias{file_setdiff.wucap_file_index}
\title{File Set Difference}
\usage{
file_setdiff(x, y)

\method{file_setdiff}{wucap_file_keys}(x, y)

\method{file_setdiff}{wucap_file_index}(x, y)
}
\arguments{
\item{x}{file keys or index}

\item{y}{file index or keys}
}
\value{
information about files in x that are not in y
}
\description{
File Set Difference
}
\examples{
\dontrun{
repo <- "~/redcap_download_files"
local_file_index <- redcap_read_file_index(repo)

## check for missing downloads
source_file_keys <-
  redcap_read_file_keys(data_dictionary, project_records)
missing_downloads <- file_setdiff(source_file_keys, local_file_index)
redcap_export_files(missing_downloads, repo, token = source_token)

## check for missing uploads
destination_file_keys <-
  redcap_read_file_keys(data_dictionary, project_records)
missing_uploads <- file_setdiff(local_file_index, destination_file_keys)
redcap_import_files(missing_uploads, repo, token = destination_token)
}
}
