% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{httr_content}
\alias{httr_content}
\title{Extract File Path or Name from a Request}
\usage{
httr_content(x, as = NULL, ...)
}
\arguments{
\item{x}{request object}

\item{as}{desired type of output: raw, text or parsed. content attempts to
automatically figure out which one is most appropriate, based on the
content-type. This function adds "file_path" and "file_name" as options.}

\item{...}{Other parameters passed to \code{\link[httr:content]{httr::content()}}.}
}
\value{
For "file_path", a path to the downloaded file. For "file_name" the
intended name of the file. For others see \code{\link[httr:content]{httr::content()}}.
}
\description{
This is an extension of \code{\link[httr:content]{httr::content()}}.
}
