% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{import_metadata}
\alias{import_metadata}
\title{Import Metadata (Data Dictionary)}
\usage{
import_metadata(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  format = c("xml", "csv", "json"),
  data,
  return_format = c("xml", "csv", "json")
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{format}{csv, json, xml \verb{[default]}}

\item{data}{The formatted data to be imported.}

\item{return_format}{csv, json, xml - specifies the format of error messages.
If you do not pass in this flag, it will select the default format for you
passed based on the 'format' flag you passed in or if no format flag was
passed in, it will default to 'xml'.}
}
\value{
\link[httr:response]{httr::response} object containing the number of fields imported.
}
\description{
This method allows you to import metadata (i.e., Data
Dictionary) into a project. Notice: Because of this method's destructive
nature, it is only available for use for projects in Development status.
}
\note{
To use this method, you must have API Import/Update privileges \emph{and}
Project Design/Setup privileges in the project.
}
\examples{
\dontrun{
import_metadata(token = my_token, format = "csv", data = "")
}
}
