% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{optional_argument}
\alias{optional_argument}
\title{Optional Argument}
\usage{
optional_argument(x, d = NULL, v = checkmate::assert_character)
}
\arguments{
\item{x}{value to process}

\item{d}{default value}

\item{v}{validation function}
}
\value{
either x or the default value after passing validation
}
\description{
Utility function to set a default value if missing or null and
to run a validation test.
}
