% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_content_type}
\alias{parse_content_type}
\title{Parse Content Type}
\usage{
parse_content_type(content_type)
}
\arguments{
\item{content_type}{content-type header from REDCap to be parsed}
}
\value{
list containing type, name, and charset
}
\description{
Parse Content Type
}
