% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{project_xml_filename}
\alias{project_xml_filename}
\title{Generate Project XML Filename}
\usage{
project_xml_filename(project_title)
}
\arguments{
\item{project_title}{name of the project}
}
\value{
a character representation of what REDCap would named the xml file
if it had been exported via the web interface
}
\description{
Generate Project XML Filename
}
