% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_file_metafile}
\alias{read_file_metafile}
\title{Read File Metafile}
\usage{
read_file_metafile(
  repo,
  metafile = c("_redcap_file_index.csv", "_redcap_download_file_log.csv",
    "_redcap_upload_file_log.csv"),
  metaclass = c("wucap_file_index", "wucap_download_file_log", "wucap_upload_file_log")
)
}
\arguments{
\item{repo}{directory containing the downloaded files and listed metafile}

\item{metafile}{the metafile to read}

\item{metaclass}{the class of metafile}
}
\value{
the contents of the metafile
}
\description{
Read File Metafile
}
