% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_project_xml}
\alias{read_project_xml}
\title{Read Project XML}
\usage{
read_project_xml(odm)
}
\arguments{
\item{odm}{path to a file}
}
\value{
collapsed character representation of the project xml file
}
\description{
Read Project XML
}
