% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{redcap_api_endpoints}
\alias{redcap_api_endpoints}
\title{REDCap API Endpoints}
\format{
An object of class \code{list} of length 3.
}
\usage{
redcap_api_endpoints
}
\description{
Constants for various API endpoints
}
\keyword{datasets}
