% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{redcap_array}
\alias{redcap_array}
\title{REDCap Array}
\usage{
redcap_array(name, values)
}
\arguments{
\item{name}{name of the array}

\item{values}{values to be included in the array}
}
\value{
an object of class "redcap_array"
}
\description{
REDCap Array
}
\examples{
## records array for three records from static
redcap_array("records", c(16227, 16342, 16419))

## fields array for two fields from static
redcap_array("fields", c("chip1_install_date", "chip2_install_date"))

## events array for one event from mother
redcap_array("events", "baseline_arm_1")
}
