% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{redcap_create_project2}
\alias{redcap_create_project2}
\title{Create Project}
\usage{
redcap_create_project2(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  project_title,
  purpose = c("Practice/Just for fun", "Other", "Research", "Quality Improvement",
    "Operational Support"),
  purpose_other,
  project_notes,
  is_longitudinal = FALSE,
  surveys_enabled = FALSE,
  record_autonumbering_enabled = FALSE,
  return_format = c("xml", "csv", "json"),
  odm
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The Super API Token specific to a user}

\item{project_title}{(required) title of the project}

\item{purpose}{(required) purpose of the project}

\item{purpose_other}{(required if purpose is "Other") custom purpose}

\item{project_notes}{(optional) notes describing the project}

\item{is_longitudinal}{is the project to be longitudinal}

\item{surveys_enabled}{is the project to contain surveys}

\item{record_autonumbering_enabled}{should the records be autonumbered}

\item{return_format}{csv, json, xml - specifies the format of error messages.
If you do not pass in this flag, it will select the default format for you
passed based on the 'format' flag you passed in or if no format flag was
passed in, it will default to 'xml'.}

\item{odm}{path to file containing an XML string in CDISC ODM XML format that
contains project metadata (fields, forms, events, arms) and might optionally
contain data to be imported as well. The XML contained in this parameter
can come from a REDCap Project XML export file from REDCap itself, or may
come from another system that is capable of exporting projects and data in
CDISC ODM format. If the 'odm' parameter is included in the API request, it
will use the XML to import its contents into the newly created project.
This will allow you not only to create the project with the API request,
but also to import all fields, forms, and project attributes (and events
and arms, if longitudinal) as well as record data all at the same time.}
}
\value{
an httr response object containing a 32-character project-level API
token (associated with both the project and user creating the project).
}
\description{
This method allows you to create a new REDCap project. A
64-character Super API Token is required for this method (as opposed to
project-level API methods that require a regular 32-character token
associated with the project-user). In the API request, you must minimally
provide the project attributes 'project_title' and 'purpose' (with numerical
value 0=Practice/Just for fun, 1=Other, 2=Research, 3=Quality Improvement,
4=Operational Support) when creating a project.

When a project is created with this method, the project will automatically be
given all the project-level defaults just as if you created a new empty
project via the web user interface, such as a automatically creating a
single data collection instrument seeded with a single Record ID field and
Form Status field, as well as (for longitudinal projects) one arm with one
event. And if you intend to create your own arms or events immediately after
creating the project, it is recommended that you utilize the override=1
parameter in the 'Import Arms' or 'Import Events' method, respectively, so
that the default arm and event are removed when you add your own. Also, the
user creating the project will automatically be added to the project as a
user with full user privileges and a project-level API token, which could
then be used for subsequent project-level API requests.

NOTE: Only users with Super API Tokens can utilize this method. Users can
only be granted a super token by a REDCap administrator (using the API Tokens
page in the REDCap Control Center). Please be advised that users with a
Super API Token can create new REDCap projects via the API without any
approval needed by a REDCap administrator. If you are interested in
obtaining a super token, please contact your local REDCap administrator.
}
\examples{
\dontrun{
redcap_create_project_from_odm(
  token = my_super_token,
  project_title = "redcap_create_project_from_odm test",
  purpose = "Practice/Just for fun",
  odm = "odm.xml"
) \%>\%
  httr::content(r, as = "text") -> my_project_token
}
}
