% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{redcap_dag_endpoints}
\alias{redcap_dag_endpoints}
\title{REDCap Data Access Group Endpoints}
\format{
An object of class \code{list} of length 3.
}
\usage{
redcap_dag_endpoints
}
\description{
Constants for various DAG endpoints
}
\keyword{datasets}
