% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{redcap_delete_records}
\alias{redcap_delete_records}
\title{Delete Records}
\usage{
redcap_delete_records(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  records,
  arm
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{records}{an array of record names specifying specific records you wish
to delete}

\item{arm}{the arm number of the arm in which the record(s) should be
deleted. (This can only be used if the project is longitudinal with more than
one arm.) NOTE: If the arm parameter is not provided, the specified records
will be deleted from all arms in which they exist. Whereas, if arm is
provided, they will only be deleted from the specified arm.}
}
\value{
\link[httr:response]{httr::response} object containing the number of records deleted.
}
\description{
Delete Records
}
\examples{
\dontrun{
## delete two records from static
token <- REDCapR::retrieve_credential_local("~/.REDCapR", 7842)$token
redcap_delete_records(
  token,
  records = redcap_array("records", c(16227, 16342))
)
}
}
