% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{redcap_export_field_names}
\alias{redcap_export_field_names}
\title{Export List of Export Field Names}
\usage{
redcap_export_field_names(
  redcap_uri = "https://redcap.wustl.edu/redcap/api/",
  token,
  format = c("xml", "csv", "json"),
  field,
  return_format = c("xml", "csv", "json")
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{format}{csv, json, xml [default]}

\item{field}{A field's variable name. By default, all fields are returned,
but if field is provided, then it will only the export field name(s) for
that field. If the field name provided is invalid, it will return an error.}

\item{return_format}{csv, json, xml - specifies the format of error messages.
If you do not pass in this flag, it will select the default format for you
passed based on the 'format' flag you passed in or if no format flag was
passed in, it will default to 'xml'.}
}
\value{
httr::response() object containing Returns a list of the
export/import-specific version of field names for all fields (or for one
field, if desired) in a project in the format specified and ordered by their
field order . The list that is returned will contain the three following
attributes for each field/choice: 'original_field_name', 'choice_value',
and 'export_field_name'. The choice_value attribute represents the raw
coded value for a checkbox choice. For non-checkbox fields, the choice_value
attribute will always be blank/empty. The export_field_name attribute
represents the export/import-specific version of that field name.
}
\description{
This method returns a list of the export/import-specific
version of field names for all fields (or for one field, if desired) in a
project. This is mostly used for checkbox fields because during data exports
and data imports, checkbox fields have a different variable name used than
the exact one defined for them in the Online Designer and Data Dictionary,
in which *each checkbox option* gets represented as its own export field
name in the following format: field_name + triple underscore + converted
coded value for the choice. For non-checkbox fields, the export field name
will be exactly the same as the original field name. Note: The following
field types will be automatically removed from the list returned by this
method since they cannot be utilized during the data import process: 'calc',
'file', and 'descriptive'.

The list that is returned will contain the three following attributes for
each field/choice: 'original_field_name', 'choice_value', and
'export_field_name'. The choice_value attribute represents the raw coded
value for a checkbox choice. For non-checkbox fields, the choice_value
attribute will always be blank/empty. The export_field_name attribute
represents the export/import-specific version of that field name.
}
\note{
To use this method, you must have API Export privileges in the project.
}
\examples{
\dontrun{
redcap_export_field_names(
  token = my_token,
  format = "csv",
  field = "checkboxes"
 ) \%>\%
 httr::content(
   as = "text",
   encoding = "UTF-8",
   content_type = "text/csv"
 ) \%>\%
 readr::read_csv() \%>\%
 dplyr::pull(export_field_name)
}
}
