% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{redcap_export_file}
\alias{redcap_export_file}
\title{Export a File}
\usage{
redcap_export_file(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  record,
  field,
  event,
  repeat_instance,
  return_format = c("xml", "csv", "json")
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{record}{the record ID}

\item{field}{the name of the field that contains the file}

\item{event}{the unique event name - only for longitudinal projects}

\item{repeat_instance}{(only for projects with repeating instruments/events)
The repeat instance number of the repeating event (if longitudinal) or the
repeating instrument (if classic or longitudinal). Default value is '1'.}

\item{return_format}{csv, json, xml - specifies the format of error messages.
If you do not pass in this flag, it will select the default format for you
passed based on the 'format' flag you passed in or if no format flag was
passed in, it will default to 'xml'.}
}
\value{
httr::response() object containing the path to the file on disk
}
\description{
This method allows you to download a document that has been
attached to an individual record for a File Upload field. Please note that
this method may also be used for Signature fields (i.e. File Upload fields
with 'signature' validation type).
}
\note{
Please be aware that Data Export user rights will be applied to this
API request. For example, if you have 'No Access' data export rights in the
project, then the API file export will fail and return an error. And if you
have 'De-Identified' or 'Remove all tagged Identifier fields' data export
rights, then the API file export will fail and return an error *only if* the
File Upload field has been tagged as an Identifier field. To make sure that
your API request does not return an error, you should have 'Full Data Set'
export rights in the project.
}
\examples{
\dontrun{
redcap_export_file(
  redcap_uri =
    "https://redcap.wustl.edu/redcap/srvrs/prod_v3_1_0_001/redcap/api/",
  token = my_project_token,
  record = 1,
  field = "file_field_name"
)
}
}
