% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{redcap_import_dags}
\alias{redcap_import_dags}
\title{Import DAGs}
\usage{
redcap_import_dags(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  format = c("xml", "csv", "json"),
  data,
  return_format = c("xml", "csv", "json")
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{format}{csv, json, xml \verb{[default]}}

\item{data}{Contains the attributes 'data_access_group_name'
(referring to the group name) and 'unique_group_name'
(referring to the auto-generated unique group name) of each DAG to be
created/modified, in which they are provided in the specified format.
\enumerate{
\item JSON Example:
\verb{[\{"data_access_group_name":"CA Site","unique_group_name":"ca_site"\} \{"data_access_group_name":"FL Site","unique_group_name":"fl_site"\}, \{"data_access_group_name":"New Site","unique_group_name":""\}]}
\item CSV Example:
data_access_group_name,unique_group_name
"CA Site",ca_site
"FL Site",fl_site
"New Site",
#' XML Example:
\out{
<?xml version="1.0" encoding="UTF-8" ?>
<dags>
  <item>
  <data_access_group_name>CA Site</data_access_group_name>
  <unique_group_name>ca_site</unique_group_name>
  </item>
  <item>
  <data_access_group_name>FL Site</data_access_group_name>
  <unique_group_name>fl_site</unique_group_name>
  </item>
  <item>
  <data_access_group_name>New Site</data_access_group_name>
  <unique_group_name></unique_group_name>
  </item>
</dags>
}
}}

\item{return_format}{csv, json, xml - specifies the format of error messages.
If you do not pass in this flag, it will select the default format for you
passed based on the 'format' flag you passed in or if no format flag was
passed in, it will default to 'xml'.}
}
\value{
\link[httr:response]{httr::response} object containing number of DAGs added or updated
}
\description{
This method allows you to import new DAGs (Data Access Groups)
into a project or update the group name of any existing DAGs.
}
\note{
DAGs can be renamed by simply changing the group name
(data_access_group_name). DAG can be created by providing group name value
while unique group name should be set to blank. To use this method, you must
have API Import/Update privileges in the project.
}
