% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{redcap_import_project_info}
\alias{redcap_import_project_info}
\title{Import Project Information}
\usage{
redcap_import_project_info(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  format = c("xml", "csv", "json"),
  data
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The Super API Token specific to a user}

\item{format}{format of the data argument}

\item{data}{Contains some or all of the attributes from Export Project
Information in the same data format as in the export. These attributes will
change the project information. Attributes for the project in the format
specified. For any values that are boolean, they should be represented as
either a '0' (no/false) or '1' (yes/true). The following project attributes
can be udpated: project_title, project_language, purpose, purpose_other,
project_notes, custom_record_label, secondary_unique_field, is_longitudinal,
surveys_enabled, scheduling_enabled, record_autonumbering_enabled,
randomization_enabled, project_irb_number, project_grant_number,
project_pi_firstname, project_pi_lastname, display_today_now_button}
}
\value{
Returns the number of values accepted to be updated in the project
settings (including values which remained the same before and after the
import).
}
\description{
Import Project Information
}
\examples{
\dontrun{
redcap_export_project_info(token = my_src_token) \%>\%
  httr::content() \%>\%
  as.character() -> xml_string_of_source_info

redcap_import_project_info(
  token = my_dst_token,
  data = xml_string_of_source_info
) \%>\%
  httr::content(as = "text")
}
}
