% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_migration_app2.R
\name{redcap_project_migration_app2}
\alias{redcap_project_migration_app2}
\title{Project Migration App}
\usage{
redcap_project_migration_app2(
  redcap_uri_src = redcap_api_endpoints$prod$v7,
  redcap_uri_dst = redcap_api_endpoints$prod$latest,
  token_src = "",
  token_dst = "",
  redcap_dag_uri_src = redcap_dag_endpoints$prod$v7,
  redcap_dag_uri_dst = redcap_dag_endpoints$prod$latest
)
}
\arguments{
\item{redcap_uri_src}{The URI (uniform resource identifier) of the source
REDCap project.}

\item{redcap_uri_dst}{The URI (uniform resource identifier) of the
destination REDCap project.}

\item{token_src}{The API token specific to your source REDCap project and
username (each token is unique to each user for each project). See the
section on the left-hand menu for obtaining a token for a given project.}

\item{token_dst}{The API token specific to your destination REDCap project
and username (each token is unique to each user for each project). See the
section on the left-hand menu for obtaining a token for a given project.}

\item{redcap_dag_uri_src}{The base URI for the source DAG page}

\item{redcap_dag_uri_dst}{The base URI for the destination DAG page}
}
\description{
This shiny app migrates most project settings from one REDCap
instance to another by using API calls. It was written to help transfer
projects on a v7 instance to a v10 instance where version inconsistencies as
well as technical and labor issues prevented a simple CDISC ODM XML port.
}
