% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{redcap_read_download_file_log}
\alias{redcap_read_download_file_log}
\title{Read Download File Log}
\usage{
redcap_read_download_file_log(repo)
}
\arguments{
\item{repo}{directory containing the downloaded files and download file log.}
}
\value{
the contents of the download file log
}
\description{
Read Download File Log
}
\examples{
\dontrun{
file_download_log <- redcap_read_download_file_log("~/redcap_download_files")
}
}
