% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{redcap_read_file_index}
\alias{redcap_read_file_index}
\title{Read File Index}
\usage{
redcap_read_file_index(repo)
}
\arguments{
\item{repo}{directory containing the downloaded files and index file.}
}
\value{
the contents of the index file
}
\description{
Read File Index
}
\examples{
\dontrun{
file_index <- redcap_read_file_index("~/redcap_download_files")
}
}
