% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{redcap_read_file_keys}
\alias{redcap_read_file_keys}
\title{Read File Keys}
\usage{
redcap_read_file_keys(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  data_dictionary,
  project_records,
  field_type = c("file", "signature", "both")
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{data_dictionary}{Project metadata (i.e., Data Dictionary values).}

\item{project_records}{Project records.}

\item{field_type}{Type of fields in which to search for files.}
}
\value{
tibble containing file keys
}
\description{
Read File Keys
}
\examples{
\dontrun{
file_keys <-
  redcap_read_file_keys(
    token = my_token,
    data_dictionary = data_dictionary,
    project_records = project_records,
    field_type = "both"
  )
}
}
