% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{redcap_reset_completion_flags}
\alias{redcap_reset_completion_flags}
\title{Reset Completion Flags}
\usage{
redcap_reset_completion_flags(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  data_dictionary,
  project_records
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{data_dictionary}{(optional) If supplied should be data object from a
call to REDCapR::redcap_metadata_read(). If not supplied the function will
call REDCapR::redcap_metadata_read().}

\item{project_records}{(optional) If supplied should be the data object from
a call to REDCapR::redcap_read(). If not supplied the function will call
REDCapR::redcap_read().}
}
\description{
When exporting records containing instrument completion flag
fields from REDCap (regardless of format), REDCap maps both
grey/no sql table record/"Incomplete (no data saved)" and
red/1/"Incomplete" to red/1/"Incomplete". Because it is difficult to
determine which records should have remained
grey/no sql table record/"Incomplete (no data saved)", this function
assumes instruments with no data present should be set to
grey/no sql table record/"Incomplete (no data saved)"
}
