% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{redcap_signature_validation_disable}
\alias{redcap_signature_validation_disable}
\title{Signature Validation Disable}
\usage{
redcap_signature_validation_disable(
  redcap_uri = redcap_api_endpoints$prod$latest,
  token,
  data_dictionary
)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{data_dictionary}{Project metadata (i.e., Data Dictionary values).}
}
\description{
Signature Validation Disable
}
\examples{
\dontrun{
redcap_signature_validation_disable(
  token = my_token,
  data_dictionary = data_dictionary
)
}
}
