% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{redcap_signature_validation_enable}
\alias{redcap_signature_validation_enable}
\title{Signature Validation Enable}
\usage{
redcap_signature_validation_enable(redcap_uri, token, data_dictionary)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{data_dictionary}{Original project metadata containing the signature
validation note (i.e., the original Data Dictionary values).}
}
\description{
Signature Validation Enable
}
\examples{
\dontrun{
redcap_signature_validation_enable(
  token = my_token,
  data_dictionary = data_dictionary
)
}
}
