% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{redcap_validate_file_index}
\alias{redcap_validate_file_index}
\title{Validate File Index}
\usage{
redcap_validate_file_index(
  repo,
  purge_rcf = c("none", "not_indexed", "not_saved", "both")
)
}
\arguments{
\item{repo}{Directory containing the downloaded files and index file.}

\item{purge_rcf}{Which REDCap files (.rcf) to purge: \code{not_indexed}: existing
local files not listed in the index file; \code{not_saved}: local files listed in
the index but not saved locally}
}
\value{
a pair of \code{\link[dplyr:reexports]{dplyr::tibble()}} listing REDCap files not saved or not
indexed
}
\description{
Validate File Index
}
\examples{
\dontrun{
## check validity
redcap_validate_file_index("my_repo")
}
}
