% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{redcap_validate_file_migration}
\alias{redcap_validate_file_migration}
\title{Validate File Migration}
\usage{
redcap_validate_file_migration(file_keys_src, file_keys_dst)
}
\arguments{
\item{file_keys_src}{file keys from the source project}

\item{file_keys_dst}{file keys from the destination project}
}
\value{
A pair of file key objects listing files in the destination project
that are not in the source project and files in the source project that are
not in the destination project.
}
\description{
Validate File Migration
}
\examples{
\dontrun{
redcap_validate_file_migration(file_keys_src, file_keys_dst)
}
}
