% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{request_file_content}
\alias{request_file_content}
\title{Extract File Path or Name from a Request}
\usage{
request_file_content(x, as = c("file_path", "file_name"))
}
\arguments{
\item{x}{request object}

\item{as}{desired type of output: file path or name}
}
\value{
For "file_path", a path to the downloaded file. For "file_name" the
intended name of the file.
}
\description{
Extract File Path or Name from a Request
}
