% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_project_xml}
\alias{write_project_xml}
\title{Write Project XML}
\usage{
write_project_xml(r, file)
}
\arguments{
\item{r}{httr response object from redcap_export_project_xml()}

\item{file}{file or connection to write to}
}
\description{
Write Project XML
}
