% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapr.R
\name{redcap_read}
\alias{redcap_read}
\title{Read records from a REDCap project in subsets, and stacks them together
before returning a dataset}
\usage{
redcap_read(project_id, path_credential = "~/.REDCapR", ...)
}
\arguments{
\item{project_id}{The ID assigned to the project withing REDCap.}

\item{path_credential}{The file path to the CSV containing the credentials.}

\item{...}{arguments passed on to
\verb{REDCapR::[redcap_read][REDCapR::redcap_read]}}
}
\description{
Read records from a REDCap project in subsets, and stacks them together
before returning a dataset
}
