% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapr.R
\name{retrieve_credential_local}
\alias{retrieve_credential_local}
\title{Read a token and other credentials from a (non-REDCap) database or file}
\usage{
retrieve_credential_local(project_id, path_credential = "~/.REDCapR", ...)
}
\arguments{
\item{project_id}{The ID assigned to the project withing REDCap.}

\item{path_credential}{The file path to the CSV containing the credentials.}

\item{...}{arguments passed on to
\verb{REDCapR::[retrieve_credential_local][REDCapR::retrieve_credential_local]}}
}
\description{
Read a token and other credentials from a (non-REDCap) database or file
}
